server/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
pillow==10.2.0
numpy==1.26.3
opencv-python==4.9.0.80
mediapipe==0.10.9
ultralytics==8.1.0
openai==1.10.0
anthropic==0.49.0
python-dotenv==1.0.0
librosa==0.10.1
pyacoustid==1.3.0
pydub==0.25.1
ffmpeg-python==0.2.0
scipy==1.12.0
